Add keypress event hook to force remaking date ticks after zoom/pan.
If <space> is pressed then date ticks are remade appropriate to the
current plot.  if '0' is pressed then date ticks are set to the
widest possible scale.  This is required if one zooms in and then
wants to pan all the way out with the Home button.  Otherwise the
zoomed-in tick scale is used which can result in a very large number
of major ticks and the appearance of hanging.